dom/base/test/jsmodules/test_asyncInlineModules.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test async inline modules</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var results = [];

  SimpleTest.waitForExplicitFinish();

  function arrayEquals(a, b) {
    if (a.length != b.length) {
      return false;
    }
    for (let i = 0; i < a.length; i++) {
      if (a[i] != b[i]) {
        return false;
      }
      return true;
    }
  }

  function testLoaded() {
    ok(arrayEquals(results.sort(), [1, 2, 3]), 'Check modules imported');
    SimpleTest.finish();
  }
</script>
<script type="module" async>
  results.push(1);
</script>
<script type="module" async>
  import "./module_simple2.mjs";
</script>
<script type="module" async>
  results.push(3);
</script>
<body onload='testLoaded()'></body>
